refactor(settings): #373 Phase 4-2 SettingsModal.tsx を 3 hook/utility に分割
Issue #373 / Phase 4-2: `components/SettingsModal.tsx` (717 行) から
section meta / focus trap / nav state を 3 つの hook/utility に分離。
move only / 機能変更なし / 公開 API は不変。

# 構成

```
components/SettingsModal.tsx              717 → ~430 行 (state 管理 + JSX shell + renderSection)
lib/settings-section-meta.tsx     新規  108 行  SectionId / ICON_PROPS / SECTION_ICON_TYPES
                                                / iconFor / FIXED_LABELS_JA / FIXED_LABELS_EN
                                                / UNTITLED_FALLBACK_* / labelOf (pure 化)
lib/hooks/use-settings-keydown.ts 新規   84 行  Issue #195: Escape + Tab focus trap
lib/hooks/use-settings-nav.ts     新規  140 行  groupsRaw / groups / activeSection 同期
```

# 公開 API の維持

- `SettingsModal` の export はそのまま
- `SettingsModalProps` interface 不変
- `iconFor` の戻り値は JSX のため `.tsx` 拡張子で配置 (caller は拡張子省略 import)
- `labelOf` は元 closure → pure 関数化 (引数: id / isJa / customAgents)
- 機能変更なし: ICON / FIXED_LABELS / UNTITLED_FALLBACK / 文言は逐字保持
- 全角/半角括弧も逐字保持 (UI 表示文字列のため)

# 検証結果

- `npm run typecheck`: ✅ green
- 機能変更なし (move only)

# Agent Team で並列分担

3 hook/utility の draft を 3 人の teammate (section-meta / settings-keydown
/ settings-nav extractor) に並列で書かせ、team-lead 側で SettingsModal.tsx
を Write で再構築 (import 切り替え + 旧定義削除 + hook 利用) を実施。

Refs #373
